+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - MINOR: add getEntryByTypeAlias when available in prod
 - MINOR: add the new bulk method to add applogs when available in prod
 - MINOR: basic robot detection with a default deviceId for them all ?
+- MINOR: have extra configuration params given go through the middleware so they are used in the factory (that will allow setting the gid when using the middleware, for instance)
 
 # 2.2.1