+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # TODO
 
 - MAJOR: remove the deprecated static method getCurrentTimeOfDayDimValue
+- MAJOR: provide the ES6 package as the default, and the ES5 one as a legacy package at appgrid/legacy
 - MINOR: basic robot detection with a default deviceId for them all ?
 - MINOR: have extra configuration params given go through the middleware so they are used in the factory (that will allow setting the gid when using the middleware, for instance)
 
@@ -13,6 +14,7 @@
 - PATCH: Changed the AppGrid REST API link to its new location
 - PATCH: doc: installing this SDK should usually be done with --save, not --save-dev
 - PATCH: the static method getCurrentTimeOfDayDimValue was deprecated, will be removed in version 3
+- PATCH: updated non-breaking dependencies
 
 # 2.2.1